+++ tests/test_lunar_sqlite3_feature.py
@@ -4,25 +4,21 @@ import sqlite3
 import unittest
 
 from borax.calendars.lunardate import LunarDate
+from borax.calendars.festivals2 import encode, decode, WrappedDate
 
-
-def adapt_lunardate(ld):
-    return ld.encode()
-
-
-sqlite3.register_adapter(LunarDate, adapt_lunardate)
-sqlite3.register_converter("lunardate", LunarDate.decode)
+sqlite3.register_adapter(WrappedDate, encode)
+sqlite3.register_converter("WrappedDate", decode)
 
 
 class Sqlite3CustomFieldTestCase(unittest.TestCase):
     def test_custom_field(self):
         con = sqlite3.connect(":memory:", detect_types=sqlite3.PARSE_DECLTYPES)
         cur = con.cursor()
-        cur.execute('CREATE TABLE member (pid INT AUTO_INCREMENT PRIMARY KEY,birthday lunardate);')
+        cur.execute('CREATE TABLE member (pid INT AUTO_INCREMENT PRIMARY KEY,birthday WrappedDate);')
         ld = LunarDate(2018, 5, 3)
-        cur.execute("INSERT INTO member(birthday) VALUES (?)", (ld,))
+        cur.execute("INSERT INTO member(birthday) VALUES (?)", (WrappedDate(ld),))
         cur.execute("SELECT pid, birthday FROM member;")
-        my_birthday = cur.fetchone()[1]
+        my_birthday = cur.fetchone()[1].lunar
         cur.close()
         con.close()
         self.assertEqual(LunarDate, type(my_birthday))